Refactor code structure for improved readability and maintainability; removed redundant sections and optimized logic flow.

+++ Codigo/Back/src/main/java/com/hortifruti/sl/hortifruti/dto/freight/DistanceFreightResponse.java
@@ -1,3 +1,5 @@
 package com.hortifruti.sl.hortifruti.dto.freight;
 
-public record DistanceFreightResponse(String distance, String duration, double freight) {}
+import java.math.BigDecimal;
+
+public record DistanceFreightResponse(String distance, String duration, BigDecimal freight) {}
