+++ Codigo/Back/src/main/java/com/hortifruti/sl/hortifruti/dto/freight/FreightConfigDTO.java
@@ -1,15 +1,17 @@
 package com.hortifruti.sl.hortifruti.dto.freight;
 
+import java.math.BigDecimal;
+
 public record FreightConfigDTO(
-    Double kmPerLiterConsumption,
-    Double fuelPrice,
-    Double maintenanceCostPerKm,
-    Double tireCostPerKm,
-    Double depreciationCostPerKm,
-    Double insuranceCostPerKm,
-    Double baseSalary,
-    Double chargesPercentage,
-    Double monthlyHoursWorked,
-    Double administrativeCostsPercentage,
-    Double marginPercentage,
-    Double fixedFee) {}
+    BigDecimal kmPerLiterConsumption,
+    BigDecimal fuelPrice,
+    BigDecimal maintenanceCostPerKm,
+    BigDecimal tireCostPerKm,
+    BigDecimal depreciationCostPerKm,
+    BigDecimal insuranceCostPerKm,
+    BigDecimal baseSalary,
+    BigDecimal chargesPercentage,
+    BigDecimal monthlyHoursWorked,
+    BigDecimal administrativeCostsPercentage,
+    BigDecimal marginPercentage,
+    BigDecimal fixedFee) {}
